basic features done

+++ app/api/password/[id]/route.ts
@@ -4,13 +4,22 @@ import { NextRequest, NextResponse } from "next/server";
 
 export async function DELETE(req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     if (!hasEnoughTimePassed(req))
-        return NextResponse.json({ message: "Requests too frequent" }, { status: 429 });
+        return NextResponse.json({ error: "Requests too frequent" }, { status: 429 });
 
     const { userId } = await auth();
     if (!userId)
         return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
 
     try {
+        const tried_Password_to_be_deleted = await prisma.triedPassword.findUnique({
+            where: {
+                user_id: userId,
+                id: (await params).id
+            }
+        });
+
+        if (!tried_Password_to_be_deleted) return NextResponse.json({ error: "No record found" }, { status: 404 });
+
         await prisma.triedPassword.delete({
             where: {
                 user_id: userId,
@@ -18,8 +27,15 @@ export async function DELETE(req: NextRequest, { params }: { params: Promise<{ i
             }
         });
 
-        return NextResponse.json({ status: 200 });
+        const deleted_password = await prisma.deletedTriedPassword.create({
+            data: {
+                user_id: userId,
+                password: tried_Password_to_be_deleted.password
+            }
+        })
+
+        return NextResponse.json({ deleted_password }, { status: 200 });
     } catch (error) {
-        return NextResponse.json({ status: "error", message: error });
+        return NextResponse.json({ error: "error", message: error }, { status: 500 });
     }
 }
